Move StatusListener to correct package
