+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/listeners/IndexingProgressListener.java
@@ -1,6 +1,5 @@
 package org.lodder.subtools.multisubdownloader.listeners;
 
-import org.lodder.subtools.multisubdownloader.gui.dialog.progress.StatusListener;
 
 public interface IndexingProgressListener extends StatusListener {
